requirements_test.txt
-r requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
pytest-homeassistant-custom-component==0.13.214
